move NBSReader into NoteBot and clean up a bit
